requirements.txt
numpy==1.22.4
opencv-contrib-python-headless
Pillow==9.1.1
requests==2.27.1
streamlit==1.10.0
streamlit_lottie==0.0.3
tensorflow==2.9.1
